requirements.txt
python-dotenv==1.0.1
duckdb==1.0.0
dbt-core==1.8.4
dbt-duckdb==1.8.1
